kpps: find the PPS device for a TTY
term exported DevicePathForTTY, which despite the name returned the TTY's own path from /proc/self/fd, and then matched it against the sysfs path attributes itself to find the PPS device.

Give kpps the whole lookup, which is where the knowledge of the kernel PPS subsystem belongs: DevicePathForTTY now takes the descriptor of a TTY with N_PPS attached and returns the /dev/ppsN fed by it. Taking the descriptor rather than a path keeps the resolution next to the match, and the caller cannot supply a name that has gone stale. term keeps the line discipline and the open, with its retry for the udev permission change.
